+++ src/Protocol/Version2.php
@@ -82,11 +82,11 @@ class Version2 implements ProtocolInterface
     }
 
     /**
-     * @return int
+     * @return positive-int
      */
     public static function getSymmetricKeyByteLength(): int
     {
-        return (int) static::SYMMETRIC_KEY_BYTES;
+        return static::SYMMETRIC_KEY_BYTES;
     }
 
     /**
