+++ src/Protocol/Version3.php
@@ -75,11 +75,11 @@ class Version3 implements ProtocolInterface
     }
 
     /**
-     * @return int
+     * @return positive-int
      */
     public static function getSymmetricKeyByteLength(): int
     {
-        return (int) static::SYMMETRIC_KEY_BYTES;
+        return static::SYMMETRIC_KEY_BYTES;
     }
 
     /**
